Compilation work.
On my Mac, this now compiles and and I can capture images.  Step one complete.

CMake and the CMAKE_MODULE_PATH are going to need work.  And until OpenCV is upgraded to version 4, Mac users will need to add the path, as per the Brew comment with opencv@2 is installed.